+++ src/vscode/treeviews/Pipelines/PowerBIPipeline.ts
@@ -5,6 +5,8 @@ import { PowerBICommandBuilder } from '../../../powerbi/CommandBuilder';
 import { PowerBIPipelineStages } from './PowerBIPipelineStages';
 import { iPowerBIPipeline } from '../../../powerbi/PipelinesAPI/_types';
 import { PowerBIPipelineOperations } from './PowerBIPipelineOperations';
+import { Helper } from '../../../helpers/Helper';
+import { ThisExtension } from '../../../ThisExtension';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class PowerBIPipeline extends PowerBIPipelineTreeItem {
@@ -27,6 +29,16 @@ export class PowerBIPipeline extends PowerBIPipelineTreeItem {
 		super.definition = value;
 	}
 
+	get _contextValue(): string {
+		let orig: string = super._contextValue;
+
+		let actions: string[] = [
+			"DELETE"
+		];
+
+		return orig + actions.join(",") + ",";
+	}
+
 	async getChildren(element?: PowerBIPipelineTreeItem): Promise<PowerBIPipelineTreeItem[]> {
 		PowerBICommandBuilder.pushQuickPickItem(this);
 
@@ -42,8 +54,17 @@ export class PowerBIPipeline extends PowerBIPipelineTreeItem {
 		return "pipelines/" + (this.id ?? this.uid);
 	}
 
-	// Dashboard-specific funtions
+	// Pipeline-specific funtions
 	public async delete(): Promise<void> {
-		//PowerBICommandBuilder.execute<iPowerBIGatewayItem>(this.apiPath, "DELETE", []);
+		let confirm: string = await PowerBICommandBuilder.showInputBox("", "Confirm deletion by typeing the Pipeline name '" + this.name + "' again.", undefined, undefined);
+		
+		if (!confirm || confirm != this.name) {
+			ThisExtension.log("Deletion of Pipeline '" + this.name + "' aborted!")
+			return;
+		}
+
+		await PowerBICommandBuilder.execute(this.apiPath, "DELETE", []);
+
+		setTimeout(() => vscode.commands.executeCommand("PowerBIPipelines.refresh", undefined, false), 1000);
 	}
 }