+++ src/vscode/treeviews/PowerBIApiTreeItem.ts
@@ -10,7 +10,7 @@ import { PowerBIApiService } from '../../powerbi/PowerBIApiService';
 import { iHandleBeingDropped } from './PowerBIApiDragAndDropController';
 
 
-export class PowerBIApiTreeItem extends vscode.TreeItem implements iPowerBIApiItem, iHandleBeingDropped {
+export class PowerBIApiTreeItem extends vscode.TreeItem implements iPowerBIApiItem {
 	protected _itemType: ApiItemType;
 	protected _id: UniqueId;
 	protected _name: string;
@@ -184,13 +184,5 @@ export class PowerBIApiTreeItem extends vscode.TreeItem implements iPowerBIApiIt
 		urlParts = urlParts.filter(x => x.length > 0)
 
 		return `v1.0/${urlParts.reverse().join("/")}/`;
-	}
-
-	// #region iHandleBeingDropped implementation
-	async handleBeingDropped(target: PowerBIApiTreeItem): Promise<void> {
-	}
-
-	get canBeDropped(): boolean {
-		return false;
 	}
 }